added an include to all of the gcc tests so I can include proper function declarations for the standard library functions they call.  also added compound literals to the parser
git-svn-id: http://k-framework.googlecode.com/svn/trunk@1413 82a86efa-2583-11df-8ea0-49ba94de29ab

+++ examples/inProgress/C/tests/lcc-tests/LCC-sort.c
@@ -61,6 +61,6 @@ int partition(a, i, j) int a[]; {
 void exchange(x, y) int *x, *y; {
 	int t;
 
-	printf("exchange(%d,%d)\n", x - xx, y - xx);
+	printf("exchange(%d,%d)\n", (int)(x - xx), (int)(y - xx));
 	t = *x; *x = *y; *y = t;
 }
